Move jumpstart helper to user test module
I think this makes sense since we're going to have a few test flows in there that utilize it.
